tests/wpt/tests/cookie-deprecation-label/cookie-deprecation-label.https.html

<!doctype html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
  promise_test(async t => {
    let label = await navigator.cookieDeprecationLabel.getValue();
    assert_equals(label, 'label_test');
  }, 'Test navigator.cookieDeprecationLabel.getValue()');
</script>
